requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.4.0
asyncer==0.0.2
attrs==23.2.0
bcrypt==4.2.0
bidict==0.23.1
bitsandbytes==0.43.2
certifi==2024.7.4
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
ctransformers==0.2.27
dataclasses-json==0.5.14
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
faiss-cpu==1.8.0.post1
fastapi==0.110.3
filelock==3.15.4
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.6.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.3
idna==3.7
importlib_metadata==8.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langsmith==0.1.93
Lazify==0.4.0
literalai==0.0.607
MarkupSafe==2.1.5
marshmallow==3.21.3
mpmath==1.3.0
multidict==6.0.5
mypy==1.11.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.6
packaging==23.2
passlib==1.7.4
pillow==10.4.0
protobuf==4.25.4
psutil==6.0.0
psycopg2==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.1
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
PyJWT==2.8.0
pypdf==4.3.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.11.3
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==72.1.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.13.1
syncer==2.0.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.26.0
urllib3==2.2.2
uvicorn==0.29.0
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
